Load CLI after bigint warning suppression
